refactor(vendingMachineMaker): 자판기 셋업 기능을 분리하기 위해서 패키지 및 클래스 생성
