+++ src/main/java/vendingmachine/Application.java
@@ -1,9 +1,12 @@
 package vendingmachine;
 
+import vendingmachine.machine.VendingMachineController;
+import vendingmachine.machine.VendingMachineMaker;
+
 public class Application {
 	public static void main(String[] args) {
-		VendingMachineController vendingMachineController = new VendingMachineController();
-		vendingMachineController.setUpInitialVendingMachine();
+		VendingMachineMaker vendingMachineMaker = new VendingMachineMaker();
+		VendingMachineController vendingMachineController = new VendingMachineController(vendingMachineMaker.setUp());
 		vendingMachineController.use();
 	}
 }
